+++ platform/src/main/java/com/eventplatform/model/GeoPosition.java
@@ -1,10 +1,9 @@
 package com.eventplatform.model;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class GeoPosition extends Entity {
-    public GeoPosition(BigInteger id, Date creationDate) {
+    public GeoPosition(int id, Date creationDate) {
         super(id, creationDate);
     }
 }
